fix: resolve dashboard LRU cache initialization error
Extract LRUCache class to separate utility module to resolve
"Cannot access 'A' before initialization" error in dashboard.

Changes:
- Create new LRUCache utility class in src/lib/utils/LRUCache.ts
- Remove duplicate inline LRUCache class from DailySummaryCard.svelte
- Add proper import for LRUCache in DailySummaryCard.svelte
- Mark cache property as readonly to satisfy ESLint
